+++ Qase.Csharp.Commons/Utils/ITestResultBuilder.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+using Qase.Csharp.Commons.Models.Domain;
+
+namespace Qase.Csharp.Commons.Utils
+{
+    public interface ITestResultBuilder
+    {
+        TestResult Build(RawTestData raw);
+    }
+}